requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.4.0
requests==2.32.3
python-multipart>=0.0.9
